fetch_strategy: fix spack.version.git_ref_lookup cycle
`for_package_version` and its helpers are package-aware. They inspect a
PackageBase instance to pick a fetch strategy, and attach a GitRefLookup
to git versions. GitRefLookup is a high-level class (it imports repo,
caches and fetch_strategy itself), so the inline import in
fetch_strategy created a cycle with the whole package machinery.

Move check_pkg_attributes, for_package_version and their helpers to
spack.package_base, next to their main consumer PackageBase.fetcher.
The AbstractRefLookup interface in spack.version.lookup remains the only
thing spack.version needs to know about lookups.
