+++ lib/spack/spack/cmd/ci.py
@@ -775,7 +775,7 @@ def validate_git_versions(
     """
     valid_commit = True
     for version in versions:
-        fetcher = spack.fetch_strategy.for_package_version(pkg, version)
+        fetcher = spack.package_base.for_package_version(pkg, version)
         assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
         with spack.stage.Stage(fetcher) as stage:
             known_commit = pkg.versions[version]["commit"]
