+++ lib/spack/spack/fetch_strategy.py
@@ -56,7 +56,6 @@ from spack.util import crypto, tty
 from spack.util.compression import decompressor_for
 from spack.util.executable import CommandNotFoundError, Executable, which
 from spack.util.filesystem import get_single_file, mkdirp, symlink, temp_cwd, working_dir
-from spack.util.string import comma_and, quote
 
 #: List of all fetch strategies, created by FetchStrategy metaclass.
 all_strategies: List[Type["FetchStrategy"]] = []
@@ -1497,208 +1496,6 @@ def from_kwargs(**kwargs) -> FetchStrategy:
             return fetcher(**kwargs)
 
     raise InvalidArgsError(**kwargs)
-
-
-def check_pkg_attributes(pkg):
-    """Find ambiguous top-level fetch attributes in a package.
-
-    Currently this only ensures that two or more VCS fetch strategies are
-    not specified at once.
-    """
-    # a single package cannot have URL attributes for multiple VCS fetch
-    # strategies *unless* they are the same attribute.
-    conflicts = set([s.url_attr for s in all_strategies if hasattr(pkg, s.url_attr)])
-
-    # URL isn't a VCS fetch method. We can use it with a VCS method.
-    conflicts -= set(["url"])
-
-    if len(conflicts) > 1:
-        raise FetcherConflict(
-            "Package %s cannot specify %s together. Pick at most one."
-            % (pkg.name, comma_and(quote(conflicts)))
-        )
-
-
-def _check_version_attributes(fetcher, pkg, version):
-    """Ensure that the fetcher for a version is not ambiguous.
-
-    This assumes that we have already determined the fetcher for the
-    specific version using ``for_package_version()``
-    """
-    all_optionals = {a for s in all_strategies for a in s.optional_attrs}
-
-    args = pkg.versions[version]
-    extra = set(args) - set(fetcher.optional_attrs) - set([fetcher.url_attr, "no_cache"])
-    extra.intersection_update(all_optionals)
-
-    if extra:
-        legal_attrs = [fetcher.url_attr] + list(fetcher.optional_attrs)
-        raise FetcherConflict(
-            "%s version '%s' has extra arguments: %s"
-            % (pkg.name, version, comma_and(quote(extra))),
-            "Valid arguments for a %s fetcher are: \n    %s"
-            % (fetcher.url_attr, comma_and(quote(legal_attrs))),
-        )
-
-
-def _extrapolate(pkg, version):
-    """Create a fetcher from an extrapolated URL for this version."""
-    try:
-        return URLFetchStrategy(url=pkg.url_for_version(version), fetch_options=pkg.fetch_options)
-    except spack.error.NoURLError:
-        raise ExtrapolationError(
-            f"Can't extrapolate a URL for version {version} because "
-            f"package {pkg.name} defines no URLs"
-        )
-
-
-def _from_merged_attrs(fetcher, pkg, version):
-    """Create a fetcher from merged package and version attributes."""
-    if fetcher.url_attr == "url":
-        mirrors = pkg.all_urls_for_version(version)
-        url = mirrors[0]
-        mirrors = mirrors[1:]
-        attrs = {fetcher.url_attr: url, "mirrors": mirrors}
-    else:
-        url = getattr(pkg, fetcher.url_attr)
-        attrs = {fetcher.url_attr: url}
-
-    attrs["fetch_options"] = pkg.fetch_options
-    attrs.update(pkg.versions[version])
-
-    if fetcher.url_attr == "git":
-        pkg_attr_list = ["submodules", "git_sparse_paths"]
-        for pkg_attr in pkg_attr_list:
-            if hasattr(pkg, pkg_attr):
-                attrs.setdefault(pkg_attr, getattr(pkg, pkg_attr))
-
-    return fetcher(**attrs)
-
-
-def for_package_version(pkg, version=None):
-    saved_versions = None
-    if version is not None:
-        saved_versions = pkg.spec.versions
-
-    try:
-        return _for_package_version(pkg, version)
-    finally:
-        if saved_versions is not None:
-            pkg.spec.versions = saved_versions
-
-
-def _for_package_version(pkg, version=None):
-    """Determine a fetch strategy based on the arguments supplied to
-    version() in the package description."""
-
-    # No-code packages have a custom fetch strategy to work around issues
-    # with resource staging.
-    if not pkg.has_code:
-        return BundleFetchStrategy()
-
-    check_pkg_attributes(pkg)
-
-    if version is not None:
-        assert not pkg.spec.concrete, "concrete specs should not pass the 'version=' argument"
-        # Specs are initialized with the universe range, if no version information is given,
-        # so here we make sure we always match the version passed as argument
-        if not isinstance(version, spack.version.StandardVersion):
-            version = spack.version.Version(version)
-
-        version_list = spack.version.VersionList()
-        version_list.add(version)
-        pkg.spec.versions = version_list
-    else:
-        version = pkg.version
-
-    # if it's a commit, we must use a GitFetchStrategy
-    commit_var = pkg.spec.variants.get("commit", None)
-    commit = commit_var.value if commit_var else None
-    tag = None
-    if isinstance(version, spack.version.GitVersion) or commit:
-        git_url = pkg.version_or_package_attr("git", version)
-        if not git_url:
-            raise spack.error.FetchError(
-                f"Cannot fetch git version for {pkg.name}. Package has no 'git' attribute"
-            )
-        if isinstance(version, spack.version.GitVersion):
-            # Populate the version with comparisons to other commits
-            from spack.version.git_ref_lookup import GitRefLookup
-
-            version.attach_lookup(GitRefLookup(pkg.name))
-
-            if not commit and version.is_commit:
-                commit = version.ref
-            version_meta_data = pkg.versions.get(version.std_version)
-        else:
-            version_meta_data = pkg.versions.get(version)
-
-        # For GitVersion, we have no way to determine whether a ref is a branch or tag
-        # Fortunately, we handle branches and tags identically, except tags are
-        # handled slightly more conservatively for older versions of git.
-        # We call all non-commit refs tags in this context, at the cost of a slight
-        # performance hit for branches on older versions of git.
-        # Branches cannot be cached, so we tell the fetcher not to cache tags/branches
-
-        # TODO(psakiev) eventually we should  only need to clone based on the commit
-
-        # commit stashed on version
-        if version_meta_data:
-            if not commit:
-                commit = version_meta_data.get("commit")
-            tag = version_meta_data.get("tag") or version_meta_data.get("branch")
-
-        kwargs = {"commit": commit, "tag": tag, "no_cache": bool(not commit)}
-        kwargs["git"] = git_url
-        kwargs["submodules"] = pkg.version_or_package_attr("submodules", version, False)
-        kwargs["git_sparse_paths"] = pkg.version_or_package_attr("git_sparse_paths", version, None)
-        kwargs["get_full_repo"] = pkg.version_or_package_attr("get_full_repo", version, False)
-
-        # if the ref_version is a known version from the package, use that version's
-        # attributes
-        ref_version = getattr(pkg.version, "ref_version", None)
-        if ref_version:
-            kwargs["git"] = pkg.version_or_package_attr("git", ref_version)
-            kwargs["submodules"] = pkg.version_or_package_attr("submodules", ref_version, False)
-
-        fetcher = GitFetchStrategy(**kwargs)
-        return fetcher
-
-    # If it's not a known version, try to extrapolate one by URL
-    if version not in pkg.versions:
-        return _extrapolate(pkg, version)
-
-    # Set package args first so version args can override them
-    args = {"fetch_options": pkg.fetch_options}
-    # Grab a dict of args out of the package version dict
-    args.update(pkg.versions[version])
-
-    # If the version specifies a `url_attr` directly, use that.
-    for fetcher in all_strategies:
-        if fetcher.url_attr in args:
-            _check_version_attributes(fetcher, pkg, version)
-            if fetcher.url_attr == "git" and hasattr(pkg, "submodules"):
-                args.setdefault("submodules", pkg.submodules)
-            return fetcher(**args)
-
-    # if a version's optional attributes imply a particular fetch
-    # strategy, and we have the `url_attr`, then use that strategy.
-    for fetcher in all_strategies:
-        if hasattr(pkg, fetcher.url_attr) or fetcher.url_attr == "url":
-            optionals = fetcher.optional_attrs
-            if optionals and any(a in args for a in optionals):
-                _check_version_attributes(fetcher, pkg, version)
-                return _from_merged_attrs(fetcher, pkg, version)
-
-    # if the optional attributes tell us nothing, then use any `url_attr`
-    # on the package.  This prefers URL vs. VCS, b/c URLFetchStrategy is
-    # defined first in this file.
-    for fetcher in all_strategies:
-        if hasattr(pkg, fetcher.url_attr):
-            _check_version_attributes(fetcher, pkg, version)
-            return _from_merged_attrs(fetcher, pkg, version)
-
-    raise InvalidArgsError(pkg, version, **args)
 
 
 def from_url_scheme(url: str, **kwargs) -> FetchStrategy:
