+++ lib/spack/spack/test/packages.py
@@ -196,7 +196,7 @@ def test_url_for_version_with_only_overrides_with_gaps(mock_packages, config):
 )
 def test_fetcher_url(spec_str, expected_type, expected_url):
     """Ensure that top-level git attribute can be used as a default."""
-    fetcher = spack.fetch_strategy.for_package_version(pkg_factory(spec_str), "1.0")
+    fetcher = spack.package_base.for_package_version(pkg_factory(spec_str), "1.0")
     assert isinstance(fetcher, expected_type)
     assert fetcher.url == expected_url
 
@@ -215,7 +215,7 @@ def test_fetcher_url(spec_str, expected_type, expected_url):
 def test_fetcher_errors(spec_str, version_str, exception_type):
     """Verify that we can't extrapolate versions for non-URL packages."""
     with pytest.raises(exception_type):
-        spack.fetch_strategy.for_package_version(pkg_factory(spec_str), version_str)
+        spack.package_base.for_package_version(pkg_factory(spec_str), version_str)
 
 
 @pytest.mark.usefixtures("mock_packages", "config")
@@ -233,9 +233,7 @@ def test_git_url_top_level_url_versions(version_str, expected_url, digest):
     # leading 62 zeros of sha256 hash
     leading_zeros = "0" * 62
 
-    fetcher = spack.fetch_strategy.for_package_version(
-        pkg_factory("git-url-top-level"), version_str
-    )
+    fetcher = spack.package_base.for_package_version(pkg_factory("git-url-top-level"), version_str)
     assert isinstance(fetcher, spack.fetch_strategy.URLFetchStrategy)
     assert fetcher.url == expected_url
     assert fetcher.digest == leading_zeros + digest
@@ -256,9 +254,7 @@ def test_git_url_top_level_url_versions(version_str, expected_url, digest):
 )
 def test_git_url_top_level_git_versions(version_str, tag, commit, branch):
     """Test git fetch strategy inference when url is specified with git."""
-    fetcher = spack.fetch_strategy.for_package_version(
-        pkg_factory("git-url-top-level"), version_str
-    )
+    fetcher = spack.package_base.for_package_version(pkg_factory("git-url-top-level"), version_str)
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert fetcher.url == "https://example.com/some/git/repo"
     assert fetcher.tag == tag
@@ -272,7 +268,7 @@ def test_git_url_top_level_git_versions(version_str, tag, commit, branch):
 def test_git_url_top_level_conflicts(version_str):
     """Test git fetch strategy inference when url is specified with git."""
     with pytest.raises(spack.fetch_strategy.FetcherConflict):
-        spack.fetch_strategy.for_package_version(pkg_factory("git-url-top-level"), version_str)
+        spack.package_base.for_package_version(pkg_factory("git-url-top-level"), version_str)
 
 
 def test_rpath_args(mutable_database):
@@ -313,7 +309,7 @@ def test_bundle_patch_directive(mock_directive_bundle, clear_directive_functions
 def test_fetch_options(version_str, digest_end, extra_options):
     """Test fetch options inference."""
     leading_zeros = "000000000000000000000000000000"
-    fetcher = spack.fetch_strategy.for_package_version(pkg_factory("fetch-options"), version_str)
+    fetcher = spack.package_base.for_package_version(pkg_factory("fetch-options"), version_str)
     assert isinstance(fetcher, spack.fetch_strategy.URLFetchStrategy)
     assert fetcher.digest == leading_zeros + digest_end
     assert fetcher.extra_options == extra_options
@@ -332,7 +328,7 @@ def test_package_can_have_sparse_checkout_properties(mock_packages, mock_fetch, 
     pkg_cls = spack.repo.PATH.get_pkg_class(spec.name)
     assert hasattr(pkg_cls, "git_sparse_paths")
 
-    fetcher = spack.fetch_strategy.for_package_version(pkg_cls(spec), "1.0")
+    fetcher = spack.package_base.for_package_version(pkg_cls(spec), "1.0")
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert hasattr(fetcher, "git_sparse_paths")
     assert fetcher.git_sparse_paths == pkg_cls.git_sparse_paths
@@ -345,7 +341,7 @@ def test_package_can_have_sparse_checkout_properties_with_commit_version(
     pkg_cls = spack.repo.PATH.get_pkg_class(spec.name)
     assert hasattr(pkg_cls, "git_sparse_paths")
 
-    fetcher = spack.fetch_strategy.for_package_version(pkg_cls(spec), "1.0")
+    fetcher = spack.package_base.for_package_version(pkg_cls(spec), "1.0")
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert hasattr(fetcher, "git_sparse_paths")
     assert fetcher.git_sparse_paths == pkg_cls.git_sparse_paths
@@ -359,7 +355,7 @@ def test_package_can_have_sparse_checkout_properties_with_gitversion(
     assert hasattr(pkg_cls, "git_sparse_paths")
 
     version = "git.foo=1.0"
-    fetcher = spack.fetch_strategy.for_package_version(pkg_cls(spec), version)
+    fetcher = spack.package_base.for_package_version(pkg_cls(spec), version)
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert hasattr(fetcher, "git_sparse_paths")
     assert fetcher.git_sparse_paths == pkg_cls.git_sparse_paths
@@ -371,11 +367,11 @@ def test_package_version_can_have_sparse_checkout_properties(
     spec = Spec("git-sparsepaths-version")
     pkg_cls = spack.repo.PATH.get_pkg_class(spec.name)
 
-    fetcher = spack.fetch_strategy.for_package_version(pkg_cls(spec), version="1.0")
+    fetcher = spack.package_base.for_package_version(pkg_cls(spec), version="1.0")
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert fetcher.git_sparse_paths == ["foo", "bar"]
 
-    fetcher = spack.fetch_strategy.for_package_version(pkg_cls(spec), version="0.9")
+    fetcher = spack.package_base.for_package_version(pkg_cls(spec), version="0.9")
     assert isinstance(fetcher, spack.fetch_strategy.GitFetchStrategy)
     assert fetcher.git_sparse_paths is None
